+++ hub3/ead/ead.go
@@ -252,6 +252,7 @@ func (dsc *Cdsc) NewNodeList(cfg *NodeConfig) (*NodeList, uint64, error) {
 		}
 		if cfg.Nodes != nil {
 			cfg.Nodes <- node
+			continue
 		}
 
 		// legacy add should not happen if there is a node channel
@@ -265,6 +266,7 @@ func (dsc *Cdsc) NewNodeList(cfg *NodeConfig) (*NodeList, uint64, error) {
 		}
 		if cfg.Nodes != nil {
 			cfg.Nodes <- node
+			continue
 		}
 
 		// legacy add should not happen if there is a node channel
@@ -278,6 +280,7 @@ func (dsc *Cdsc) NewNodeList(cfg *NodeConfig) (*NodeList, uint64, error) {
 		}
 		if cfg.Nodes != nil {
 			cfg.Nodes <- node
+			continue
 		}
 
 		// legacy add should not happen if there is a node channel