+++ hub3/fragments/v1.go
@@ -213,14 +213,25 @@ type Legacy struct {
 }
 
 // NewLegacy returns a legacy struct with default values
-func NewLegacy(indexDoc map[string]interface{}, fb *FragmentBuilder) *Legacy {
+func NewLegacy(indexDoc map[string]interface{}, fb *FragmentBuilder, recordTypes ...string) *Legacy {
 	l := &Legacy{
 		HubID:      fb.fg.Meta.GetHubID(),
-		RecordType: "mdr",
 		Spec:       fb.fg.Meta.GetSpec(),
 		OrgID:      fb.fg.Meta.GetOrgID(),
+		RecordType: "mdr",
 		Collection: fb.fg.Meta.GetSpec(),
 	}
+
+	if len(recordTypes) != 0 {
+		types := []string{}
+		for _, t := range recordTypes {
+			if t != "" {
+				types = append(types, t)
+			}
+		}
+		l.RecordType = strings.Join(types, ";")
+	}
+
 	var ok bool
 	_, ok = indexDoc["nave_geoHash"]
 	l.HasGeoHash = strconv.FormatBool(ok)
@@ -349,6 +360,7 @@ func GetNaveField(s string) r.Term {
 // GetUrns returs a list of WebResource urns
 func (fb *FragmentBuilder) GetUrns() []string {
 	var urns []string
+	// TODO(kiivihal): replace with internal graph package
 	wrs := fb.Graph.All(nil, nil, GetEDMField("WebResource"))
 	for _, t := range wrs {
 		s := strings.Trim(t.Subject.String(), "<>")
@@ -698,8 +710,10 @@ func (fb *FragmentBuilder) GetResourceLabel(t *r.Triple) (string, bool) {
 	case *r.Resource:
 		id := r.GetResourceID(t.Object)
 		label, ok := fb.ResourceLabels[id]
+
 		return label, ok
 	}
+
 	return "", false
 }
 
@@ -711,8 +725,13 @@ func (fb *FragmentBuilder) SetResourceLabels() error {
 		r.NewResource("http://www.w3.org/2004/02/skos/core#prefLabel"),
 		r.NewResource("http://xmlns.com/foaf/0.1/name"),
 	}
-	for _, label := range labels {
-		for _, t := range fb.Graph.All(nil, label, nil) {
+
+	for t := range fb.Graph.IterTriples() {
+		for _, label := range labels {
+			if !label.Equal(t.Predicate) {
+				continue
+			}
+
 			subjectID := t.GetSubjectID()
 			_, ok := fb.ResourceLabels[subjectID]
 			if ok {
@@ -721,6 +740,7 @@ func (fb *FragmentBuilder) SetResourceLabels() error {
 			fb.ResourceLabels[subjectID] = t.Object.(*r.Literal).RawValue()
 		}
 	}
+
 	return nil
 }
 
@@ -734,7 +754,7 @@ func fieldsContains(s []*IndexEntry, e *IndexEntry) bool {
 }
 
 // CreateV1IndexDoc creates a map that can me marshaled to json
-func CreateV1IndexDoc(fb *FragmentBuilder) (map[string]interface{}, error) {
+func CreateV1IndexDoc(fb *FragmentBuilder, recordTypes ...string) (map[string]interface{}, error) {
 	indexDoc := make(map[string]interface{})
 
 	// set the resourceLabels
@@ -781,7 +801,7 @@ func CreateV1IndexDoc(fb *FragmentBuilder) (map[string]interface{}, error) {
 	indexDoc["revision"] = fb.fg.Meta.GetRevision()
 	indexDoc["hubID"] = fb.fg.Meta.GetHubID()
 	indexDoc["system"] = NewSystem(indexDoc, fb)
-	indexDoc["legacy"] = NewLegacy(indexDoc, fb)
+	indexDoc["legacy"] = NewLegacy(indexDoc, fb, recordTypes...)
 
 	return indexDoc, nil
 }